Various fixes and matches
Former-commit-id: f75b37e7736bc56d8172d763a621ea54dec4333f

+++ src/musyx/runtime/hardware.c
@@ -423,7 +423,7 @@ void hwFlushStream(void* base, unsigned long offset, unsigned long bytes, unsign
 }
 
 void hwPrepareStreamBuffer() {}
-u32 hwInitStream(u32 len) { return aramAllocateStreamBuffer(len); }
+u8 hwInitStream(u32 len) { return aramAllocateStreamBuffer(len); }
 
 void hwExitStream(u8 id) { aramFreeStreamBuffer(id); }
